Convert monitors to be driven by entry_points
