+++ jupyter_server/services/kernelspecs/monitors/__init__.py
@@ -1,0 +1,1 @@
+from .watchdog_monitor import KernelSpecWatchdogMonitor  # noqa
